+++ src/main/java/com/shazam/fork/system/axmlparser/AXMLParser.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.squareup.spoon.axmlparser;
+package com.shazam.fork.system.axmlparser;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -22,6 +22,7 @@ import java.io.InputStream;
  * Parser for Android's binary xml files (axml).
  * @author Dmitry Skiba
  */
+@SuppressWarnings("ALL")
 public final class AXMLParser {
 
     /**
