+++ src/main/java/com/shazam/fork/system/axmlparser/Cast.java
@@ -13,37 +13,19 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.squareup.spoon.axmlparser;
+package com.shazam.fork.system.axmlparser;
 
 /**
  * @author Dmitry Skiba
  *
- * Helper class, used in Cast.toCharSequence.
  */
-public class CSString implements CharSequence {
+@SuppressWarnings("ALL")
+public class Cast {
 
-    public CSString(String string) {
+    public static final CharSequence toCharSequence(String string) {
         if (string==null) {
-            string="";
+            return null;
         }
-        m_string=string;
+        return new CSString(string);
     }
-
-    public int length() {
-        return m_string.length();
-    }
-
-    public char charAt(int index) {
-        return m_string.charAt(index);
-    }
-
-    public CharSequence subSequence(int start,int end) {
-        return new CSString(m_string.substring(start,end));
-    }
-
-    public String toString() {
-        return m_string;
-    }
-
-    private String m_string;
 }
